Updated to fix #5 and fix #19 We only need to support pawn promotion via some mechanism
 * Added PacketHandler
   *  Added ChessPlay as a message for updating the WoodChessBoard
TileEntity from the client.
 * Insulated WoodChessBoardBlock's onBlockActivated from a direct call
to client-side code so it's compatible with server and client
distributions
 * Added square selection in the WoodChessBoardGUI and made it use the
new ChessPlay message to update the server
 * Removed WoodChessboardTileEntity implementation of Ticking. It no
longer ticks! Updates are handled manually (and sensibly)
 * WoodChessboardTileEntity now passes 42 as a TileEntityType because
that's what Vanilla does, but I'm still not sure what this is for
 * Re-added some call to super in WoodChessboardTileEntity#getUpdateTag
for no real reason, but I should probably do it elsewhere for when I
extract a superclass
 * Added WoodChessboardTileEntity#handlUpdateTag since it's part of the
full set for handling client updates (as far as I can tell, though it
appeared ok without)
 * Added something to be called when the client is being changed by a
ChessPlay message... some validation should eventually go here and it
should eventually also be doing the validation