Added Elo ranking and fixed a minor gui bug
 * Fixed #28 added Elo ranking... you just can't play against anyone
else yet... but the `Capability` is added and when you Checkmate someone
on the gold board you'll get scored as if you played yourself... maybe I
set `K` to `30` instead of `32` because I checked against a calculator
that didn't let you put `32`?
 * Fixed #79 pawns weren't persisting their moved status across board
clones so I made a regression test and fixed it by checking side and
y-position against expected initial y-position.