It's getting warmer
Now the game doesn't crash which is good news. Now, we need to find a way that it works and doesn't crash when the player uses the chess wrench or not. Instead of using Registration.Chesswrench.get(), I used ChessWrench.isPresent. I don't know if we could use it in an another way. We have to verify that so I could port the rest of the versions ASAP